feat: middleware 삭제 및 route로 이동
